Add support for alternate emulation speed
Added support for a user-configurable alternate emulation speed
between 1 FPS and 240 FPS (180 is the default).  The default binding
for this is the TAB key; while the key is pressed the emulator runs at
the alternate speed, when released it returns to the normal speed for
the chosen system type.

Fixed calculation used in FPS display.

Added boolean option to control whether or not audio is muted when
running at anything other than the chosen system type's native speed.
This defaults to 'true'.