Add unzip support
Add support for standard ROM formats in zip files.  All files in the
top-level directory of the zip file are checked in order until a valid
ROM is found.  The name/extension of the ROM file inside the zip needs
to be a valid NES ROM extension (nes, fds, unf, unif).

Add support for automatically applying patches inside a zip file
containing a ROM.  This works similarly to the existing autopatch
support, except the patches don't need to start with the name of the
ROM file; it's assumed that these patches are meant to apply to the
one ROM inside the zip.  They are still applied in lexicographic
order, so they can be named with a numeric prefix to get them to be
applied in the desired order.  All patches must be placed in the
top-level directory of the zip file in order for them to be applied.

If a zipped ROM contains patches, these will be used instead of any
patches in the patch directory or in the ROM directory that might have
otherwise been used by the autopatcher.

The user may place any other files he/she wants into the zip.  It is
recommended that any other files aside from the ROM and patches be
placed in sub-directories inside the zip so that cxNES won't try to
load them as ROMs and slow down the loading process.  This allows the
user to include large data files without cxNES trying to load them.

Disabling autopatches will not prevent zipped patches from being
applied, since the whole reason to include patches in the zip is to
have them applied every time without actually hard-patching the ROM.