Add support for split roms and more
Added support for split roms inside zipfiles; ROM must exist in the
database for this to work, as there is no support for metadata in the
archive itself.  This was done primarily to support MAME-format
PlayChoice ROMs and Vs. System ROMs.

Romdb is now much larger, pulling in split rom checksums data from
both NesCart DB and MESS.

Split roms can contain patches as well; they work the same way as
zipped headered roms with patches, except that they all need to reside
in the same directory within the zip as the first PRG ROM file.

If a zipped split PC10 ROM has identical PRG and CHR to an NES
release, and the zip also contains the file "security.prm", make
sure that PlayChoice is the system type used instead of NES.