Fix handling of system type and input devices
Previously system type and input devices were automatically remembered
by storing them in the rom-specific config file.  Now, this is only
true if you check "Remember System Type" in the System Type submenu or
"Remember Input Devices" in the Input submenu.

If "Remember Input Devices" is checked, the chosen devices for ports
1-4, the expansion port, and the four-player type are saved and loaded
the next time you load the rom.  If not, they are all reset to "Auto",
which will either be determined by the database (if the rom is known
and you're using the database) or else through the header, if
possible.

The same goes with system type.  Unless "Remember System Type" is
checked, the system type will be "Auto" the next time you load the
rom.

Both of these new settings are also rom-specific and default to false.

Right now selection of "preferred" console type is broken, but
that should be fixed soon.