Add some homebrew ROMs to the database
There are a number of homebrew ROMs that support multiple video timings, and in
order for these to work correctly with overclocking they need to be handled
like the dual-timing Camerica roms: run for a number of frames without
overclocking enabled on reset, then switch to the desired overclocking mode.
If we don't flag them in the database then they will run faster than expected
in one way or another (music tempo, gameplay, etc).

The support for multiple timings could also be included in the ROM header, but
none of these ROMs do that.